# docs/swagger.yaml
definitions:
...
  ports.CreateNotificationInput:
    properties:
      action_url:
        type: string
      message:
        type: string
      notification_type:
        type: string
      receiver_user_id:
        type: integer
      related_entity_id:
        type: integer
      related_entity_type:
        type: string
      sender_user_id:
        type: integer
      title:
        maxLength: 255
        type: string
    required:
    - message
    - notification_type
    - receiver_user_id
    - title
    type: object
...
  ports.NotificationResponse:
    properties:
      action_url:
        type: string
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      notification_type:
        type: string
      read:
        type: boolean
      receiver:
        $ref: '#/definitions/ports.UserResponse'
      related_entity_id:
        type: integer
      related_entity_type:
        type: string
      sender:
        $ref: '#/definitions/ports.UserResponse'
      title:
        type: string
    type: object
...
  ports.UserResponse:
    properties:
      active:
        type: boolean
      contact_email:
        type: string
      contact_phone_no:
        type: string
      created_at:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      login_email:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /notifications:
    post:
      consumes:
      - application/json
      description: Creates a new notification record. Requires authentication, typically
        used by system services.
      parameters:
      - description: Notification details (ReceiverUserID, Title, Message, Type)
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/ports.CreateNotificationInput'
      produces:
      - application/json
      responses:
        "201":
          description: Notification created successfully
          schema:
            $ref: '#/definitions/ports.NotificationResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrReceiverNotFound
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Notification (Internal/System Use)
      tags:
      - notifications
...
  /users/{id}/notifications:
    get:
      description: Retrieves all notifications for the specified user. Requires authentication
        and self-management.
      parameters:
      - description: Target User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by read status (true/false)
        in: query
        name: read
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of notifications
          schema:
            items:
              $ref: '#/definitions/ports.NotificationResponse'
            type: array
        "400":
          description: Invalid user ID or query parameter
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the target user)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Notifications for User
      tags:
      - notifications
      - users
  /users/{id}/notifications/{notificationID}:
    delete:
      description: Deletes a specific notification for the user. Requires authentication
        and self-management.
      parameters:
      - description: Target User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Notification ID
        in: path
        name: notificationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Notification deleted successfully (No Content)
        "400":
          description: Invalid user/notification ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the target user)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Notification not found for this user
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete Single Notification
      tags:
      - notifications
      - users
  /users/{id}/notifications/{notificationID}/read:
    patch:
      description: Marks a specific notification as read. Requires authentication
        and self-management.
      parameters:
      - description: Target User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Notification ID
        in: path
        name: notificationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Notification marked as read
          schema:
            $ref: '#/definitions/ports.NotificationResponse'
        "400":
          description: Invalid user/notification ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the target user)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Notification not found for this user
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark Single Notification as Read
      tags:
      - notifications
      - users
  /users/{id}/notifications/read-all:
    patch:
      description: Marks all unread notifications for the specified user as read.
        Requires authentication and self-management.
      parameters:
      - description: Target User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Count of notifications marked as read
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the target user)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark All Notifications as Read
      tags:
      - notifications
      - users
...
swagger: "2.0"